+++ app/routers/players.py
@@ -1,0 +1,100 @@
+from fastapi import APIRouter, Depends, HTTPException
+from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.future import select
+from sqlalchemy import delete
+import logging
+
+from app.models import Player, Match
+from app.schemas import PlayerCreate
+from app.database import get_db
+from app.auth import is_admin
+
+app = APIRouter()
+logger = logging.getLogger(__name__)
+
+
+@app.post("/players")
+async def add_player(player: PlayerCreate, db: AsyncSession = Depends(get_db), admin=Depends(is_admin)):
+    new_player = Player(name=player.name)  # ✅ No duplicate check
+    db.add(new_player)
+    await db.commit()
+    
+    return {"message": f"Player {player.name} added successfully!", "rating": 1500, "matches": 0}
+
+@app.get("/players")
+async def get_players(db: AsyncSession = Depends(get_db)):
+    result = await db.execute(select(Player))
+    players = result.scalars().all()
+    return [{"id": p.id, "name": p.name, "rating": p.rating, "matches": p.matches} for p in players]
+
+@app.get("/players/{player_id}")
+async def get_player(player_id: int, db: AsyncSession = Depends(get_db)):
+    logger.info(f"Fetching player with ID: {player_id}")
+    print(f"Fetching player with ID: {player_id}")
+
+    try:
+        result = await db.execute(select(Player).where(Player.id == player_id))
+        player = result.scalars().first()
+
+        if not player:
+            logger.warning(f"Player {player_id} not found.")
+            raise HTTPException(status_code=404, detail="Player not found.")
+
+        # Debugging: Check if 'handedness' exists
+        if not hasattr(player, "handedness"):
+            print("DEBUG: 'handedness' attribute is missing from the Player model!")
+            raise HTTPException(status_code=500, detail="Player model does not match database schema")
+
+        logger.info(f"Player found: {player.name} (ID: {player.id})")
+
+        return {
+            "id": player.id,
+            "name": player.name,
+            "rating": player.rating,
+            "matches": player.matches if player.matches is not None else 0,
+            "handedness": player.handedness if player.handedness is not None else "Unknown",
+            "forehand_rubber": player.forehand_rubber or "Unknown",
+            "backhand_rubber": player.backhand_rubber or "Unknown",
+            "blade": player.blade or "Unknown",
+            "age": player.age if player.age is not None else "Unknown",
+            "gender": player.gender or "Unknown"
+        }
+
+    except Exception as e:
+        logger.error(f"Error fetching player {player_id}: {e}", exc_info=True)
+        raise HTTPException(status_code=500, detail="Internal Server Error")
+    
+@app.delete("/players/{player_id}")
+async def delete_player(player_id: int, db: AsyncSession = Depends(get_db)):
+    result = await db.execute(select(Player).where(Player.id == player_id))
+    player = result.scalars().first()
+
+    if not player:
+        raise HTTPException(status_code=404, detail="Player not found.")
+
+    # ✅ Delete all matches where the player was involved
+    await db.execute(delete(Match).where((Match.player1_id == player_id) | (Match.player2_id == player_id)))
+
+    # ✅ Delete the player after removing matches
+    await db.delete(player)
+    await db.commit()
+
+    return {"message": f"Player {player.name} and their matches deleted successfully."}
+
+@app.patch("/players/{player_id}")
+async def update_player(player_id: int, player_update: dict, db: AsyncSession = Depends(get_db), admin=Depends(is_admin)):
+    stmt = select(Player).where(Player.id == player_id)
+    result = await db.execute(stmt)
+    player = result.scalars().first()
+
+    if not player:
+        raise HTTPException(status_code=404, detail="Player not found")
+
+    for key, value in player_update.items():
+        if hasattr(player, key) and value is not None:
+            setattr(player, key, value)
+
+    await db.commit()
+    await db.refresh(player)
+    return player
+
